requirements.txt
streamlit
transformers
torch
opencv-python
Pillow
sqlalchemy
python-dotenv
pytesseract
openai
torchvision
requests
